feat: add read and write to auth

+++ backend/src/item.rs
@@ -4,7 +4,7 @@ use actix_web::{delete, get, patch, post, web, HttpResponse};
 use serde::{Deserialize, Serialize};
 use sqlx::{Pool, Postgres};
 
-use crate::{auth::check_auth, error::Error};
+use crate::{auth::{check_auth, Permission}, error::Error};
 
 #[derive(Serialize, Deserialize, Debug)]
 pub(crate) struct ItemGetResponse {
@@ -60,7 +60,7 @@ pub(crate) async fn get_item(
     let club = club.as_ref();
     let mut pool = pool.get_ref().begin().await?;
 
-    check_auth(id, session, club).await?;
+    check_auth(&id, &session, club).await?;
 
     let items = if let ItemGetQuery {
         column: Some(column),
@@ -116,7 +116,9 @@ pub(crate) async fn add_item(
     let club = club.as_ref();
     let mut pool = pool.get_ref().begin().await?;
 
-    check_auth(id, session, club).await?;
+    if !matches!(check_auth(&id, &session, club).await?, Permission::Write) {
+        return Err(Error::Unauthorized);
+    }
 
     let item: ItemAddRequest = serde_json::from_str(&body)?;
 
@@ -173,7 +175,9 @@ pub(crate) async fn update_item(
     let club = club.as_ref();
     let mut pool = pool.get_ref().begin().await?;
 
-    check_auth(id, session, club).await?;
+    if !matches!(check_auth(&id, &session, club).await?, Permission::Write) {
+        return Err(Error::Unauthorized);
+    }
 
     let item: ItemUpdateRequest = serde_json::from_str(&body)?;
 
@@ -229,7 +233,9 @@ pub(crate) async fn delete_item(
     let club = club.as_ref();
     let mut pool = pool.get_ref().begin().await?;
 
-    check_auth(id, session, club).await?;
+    if !matches!(check_auth(&id, &session, club).await?, Permission::Write) {
+        return Err(Error::Unauthorized);
+    }
 
     sqlx::query!(
         "DELETE FROM items WHERE id = $1 AND club = $2",
